auth config

+++ src/main/java/org/fstt/dao/CommandeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommandeRepository extends JpaRepository<Commande, Integer> {
+public interface CommandeRepository extends JpaRepository<Commande, Long> {
 
 }
