+++ src/main/java/org/fstt/dao/LigneCommandeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LigneCommandeRepository extends JpaRepository<LigneCommande, Integer> {
+public interface LigneCommandeRepository extends JpaRepository<LigneCommande, Long> {
 
 }
