+++ src/main/java/org/fstt/dao/LigneLivraisonRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LigneLivraisonRepository extends JpaRepository<LigneLivraison, Integer> {
+public interface LigneLivraisonRepository extends JpaRepository<LigneLivraison, Long> {
 
 }
