+++ src/main/java/org/fstt/metier/ArticleMetier.java
@@ -7,11 +7,11 @@ import org.fstt.entities.Article;
 public interface ArticleMetier {
 	public Article saveArticle(Article article);
 	
-	public Article updateArticle(Article article, Integer id);
+	public Article updateArticle(Article article, Long id);
 	
-	public Article getArticle(Integer id);
+	public Article getArticle(Long id);
 	
-	public void deleteArticle(Integer id);
+	public void deleteArticle(Long id);
 	
 	public List<Article> listArticle();
 }
