+++ src/main/java/org/fstt/metier/ClientMetier.java
@@ -7,11 +7,11 @@ import org.fstt.entities.Client;
 public interface ClientMetier {
 	public Client saveClient(Client client);
 	
-	public Client updateClient(Client client, Integer id);
+	public Client updateClient(Client client, Long id);
 	
-	public Client getClient(Integer id);
+	public Client getClient(Long id);
 	
-	public void deleteClient(Integer id);
+	public void deleteClient(Long id);
 	
 	public List<Client> listClient();
