+++ src/main/java/org/fstt/metier/CommandeMetier.java
@@ -5,13 +5,13 @@ import java.util.List;
 import org.fstt.entities.Commande;
 
 public interface CommandeMetier {
-	public Commande saveCommande(Commande commande, Integer id);
+	public Commande saveCommande(Commande commande, Long id);
 	
-	public Commande updateCommande(Commande commande, Integer id);
+	public Commande updateCommande(Commande commande, Long id);
 	
-	public Commande getCommande(Integer id);
+	public Commande getCommande(Long id);
 	
-	public void deleteCommande(Integer id);
+	public void deleteCommande(Long id);
 	
 	public List<Commande> listCommande();
