+++ src/main/java/org/fstt/metier/FournisseurMetier.java
@@ -7,11 +7,11 @@ import org.fstt.entities.Fournisseur;
 public interface FournisseurMetier {
 	public Fournisseur saveFournisseur(Fournisseur fournisseur);
 	
-	public Fournisseur updateFournisseur(Fournisseur fournisseur, Integer id);
+	public Fournisseur updateFournisseur(Fournisseur fournisseur, Long id);
 	
-	public Fournisseur getFournisseur(Integer id);
+	public Fournisseur getFournisseur(Long id);
 	
-	public void deleteFournisseur(Integer id);
+	public void deleteFournisseur(Long id);
 	
 	public List<Fournisseur> listFournisseur();
 }
