+++ src/main/java/org/fstt/service/FournisseurRestService.java
@@ -22,17 +22,17 @@ public class FournisseurRestService {
 	}
 	
 	@RequestMapping(value = "/updateFournisseur/{id}", method = RequestMethod.PUT)
-	public Fournisseur updateFournisseur(@RequestBody Fournisseur fournisseur, @PathVariable Integer id) {
+	public Fournisseur updateFournisseur(@RequestBody Fournisseur fournisseur, @PathVariable Long id) {
 		return fournisseurMetier.updateFournisseur(fournisseur, id);
 	}
 	
 	@RequestMapping(value = "/getFournisseur/{id}", method = RequestMethod.GET)
-	public Fournisseur getFournisseur(@PathVariable Integer id) {
+	public Fournisseur getFournisseur(@PathVariable Long id) {
 		return fournisseurMetier.getFournisseur(id);
 	}
 	
 	@RequestMapping(value = "/deleteFournisseur/{id}", method = RequestMethod.DELETE)
-	public void deleteFournisseur(@PathVariable Integer id) {
+	public void deleteFournisseur(@PathVariable Long id) {
 		fournisseurMetier.deleteFournisseur(id);
 	}
